requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# app/api/image/views.py: 15,16,17
# app/api/support/views.py: 14,15,16,17,18
# app/api/task/views.py: 13
# app/common/base.py: 14,15,16
# app/common/decorator.py: 8,9,10,11,12
# app/config.py: 10
# app/models.py: 16
# app/views/admin.py: 11,12
# app/views/api/wx.py: 11,12
# app/views/article.py: 8,9,10
# app/views/video.py: 10,11,12
# run.py: 11,12
# run_admin.py: 16,17
Flask == 0.12

# app/run.py: 18
Flask_Restless == 0.17.0

# app/config.py: 11
Flask_SQLAlchemy == 2.1

# app/common/base.py: 22
PyMySQL == 0.7.9

# app/api/image/service.py: 15
# app/common/base.py: 4,18
# app/models.py: 17
SQLAlchemy == 1.1.5

# app/config.py: 12,13
Werkzeug == 0.11.15

# app/crawler/crawler.py: 13,14
# tests/test_mao.py: 1
beautifulsoup4 == 4.6.0

# app/common/elasticsearch.py: 8,9,10,11
elasticsearch == 5.2.0

# app/common/aliyun.py: 7
# app/common/qiniu.py: 7
oss2 == 2.3.2

# app/common/security.py: 11
pycrypto == 2.6.1

# app/api/image/service.py: 8
# app/common/aliyun.py: 8
# app/common/qiniu.py: 8
# app/common/third/juhe.py: 7
# app/common/third/wx.py: 14
# app/crawler/crawler.py: 18
# tests/test_aliyun.py: 15
# tests/test_db.py: 11
# tests/test_mao.py: 2
requests == 2.13.0

# app/common/third/wx.py: 15
# tests/test_xml.py: 1
xmltodict == 0.11.0
